requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
attrs==25.3.0
bcc==0.18.0
beautifulsoup4==4.13.3
blinker==1.8.2
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-spectacular==0.28.0
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.0.3
fonttools==4.54.1
gyp==0.1
h11==0.13.0
hidpidaemon==18.4.6
httplib2==0.20.2
httptools==0.2.0
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
matplotlib==3.9.2
mne==1.8.0
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.0
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
pop-transition==1.1.2
protobuf==3.12.4
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.2
pydantic_core==2.33.1
pydbus==0.6.0
pyEDFlib==0.1.38
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.11.3
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.0
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.36.2
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
rpds-py==0.24.0
scipy==1.14.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.46.1
systemd-python==234
tqdm==4.66.5
typing-inspection==0.4.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.15.0
uvloop==0.21.0
wadllib==1.3.6
watchgod==0.8.2
websockets==15.0.1
Werkzeug==3.0.4
wsproto==1.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
